requirements.txt
maco==1.1.0
pycryptodome==3.20.0
pycryptodomex==3.20.0
